backend/requirements.txt
# backend/requirements.txt

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg==3.1.19
python-dotenv==1.0.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
PyJWT==2.8.0